refactor : 시간 계산 함수들 따로 분리
- src/utils/ 에서 dateFormatter.js 파일로 관리
